Remove duplicate code in pipeline run table toolbar
