+++ frontend/src/concepts/pipelines/content/compareRuns/CompareRunTableToolbar.tsx
@@ -1,35 +1,17 @@
 import * as React from 'react';
-import { Button, TextInput, ToolbarItem } from '@patternfly/react-core';
+import { Button, ToolbarItem } from '@patternfly/react-core';
 import { useNavigate, useParams } from 'react-router';
-import PipelineFilterBar from '~/concepts/pipelines/content/tables/PipelineFilterBar';
-import SimpleSelect from '~/components/SimpleSelect';
 import { FilterOptions } from '~/concepts/pipelines/content/tables/usePipelineFilter';
-import ExperimentSearchInput from '~/concepts/pipelines/content/tables/ExperimentSearchInput';
-import { RuntimeStateKF, runtimeStateLabels } from '~/concepts/pipelines/kfTypes';
-import DashboardDatePicker from '~/components/DashboardDatePicker';
-import PipelineVersionSelect from '~/concepts/pipelines/content/pipelineSelector/CustomPipelineVersionSelect';
-import { PipelineRunVersionsContext } from '~/pages/pipelines/global/runs/PipelineRunVersionsContext';
 import { experimentsBaseRoute, experimentsManageCompareRunsRoute } from '~/routes';
 import { useCompareRuns } from '~/concepts/pipelines/content/compareRuns/CompareRunsContext';
-
-export type FilterProps = Pick<
-  React.ComponentProps<typeof PipelineFilterBar>,
-  'filterData' | 'onFilterUpdate'
->;
+import PipelineRunTableToolbarBase, {
+  FilterProps,
+} from '~/concepts/pipelines/content/tables/pipelineRun/PipelineRunTableToolbarBase';
 
 const CompareRunTableToolbar: React.FC<FilterProps> = ({ ...toolbarProps }) => {
-  const { versions } = React.useContext(PipelineRunVersionsContext);
   const { runs } = useCompareRuns();
   const navigate = useNavigate();
   const { namespace, experimentId } = useParams();
-  /* eslint-disable @typescript-eslint/no-unused-vars */
-  const {
-    [RuntimeStateKF.RUNTIME_STATE_UNSPECIFIED]: unspecifiedState,
-    [RuntimeStateKF.PAUSED]: pausedState,
-    [RuntimeStateKF.CANCELED]: cancelledState,
-    ...statusRuntimeStates
-  } = runtimeStateLabels;
-  /* eslint-enable @typescript-eslint/no-unused-vars */
 
   const options = React.useMemo(
     () => ({
@@ -48,60 +30,7 @@ const CompareRunTableToolbar: React.FC<FilterProps> = ({ ...toolbarProps }) => {
   }
 
   return (
-    <PipelineFilterBar
-      {...toolbarProps}
-      filterOptions={options}
-      filterOptionRenders={{
-        [FilterOptions.NAME]: ({ onChange, ...props }) => (
-          <TextInput
-            {...props}
-            data-testid="search-for-run-name"
-            aria-label="Search for a run name"
-            placeholder="Search..."
-            onChange={(_event, value) => onChange(value)}
-          />
-        ),
-        [FilterOptions.EXPERIMENT]: ({ onChange, value, label }) => (
-          <ExperimentSearchInput
-            onChange={(data) => onChange(data?.value, data?.label)}
-            selected={value && label ? { value, label } : undefined}
-          />
-        ),
-        [FilterOptions.PIPELINE_VERSION]: ({ onChange, label }) => (
-          <PipelineVersionSelect
-            versions={versions}
-            selection={label}
-            onSelect={(version) => onChange(version.pipeline_version_id, version.display_name)}
-          />
-        ),
-        [FilterOptions.CREATED_AT]: ({ onChange, ...props }) => (
-          <DashboardDatePicker
-            {...props}
-            hideError
-            aria-label="Select a start date"
-            onChange={(_, value, date) => {
-              if (date || !value) {
-                onChange(value);
-              }
-            }}
-          />
-        ),
-        [FilterOptions.STATUS]: ({ value, onChange, ...props }) => (
-          <SimpleSelect
-            {...props}
-            value={value}
-            aria-label="Select a status"
-            options={Object.values(statusRuntimeStates).map((v) => ({
-              key: v,
-              label: v,
-            }))}
-            toggleLabel={value}
-            onChange={(selection) => onChange(selection)}
-            dataTestId="runtime-status-dropdown"
-          />
-        ),
-      }}
-    >
+    <PipelineRunTableToolbarBase {...toolbarProps} filterOptions={options}>
       <ToolbarItem>
         <Button
           variant="primary"
@@ -118,7 +47,7 @@ const CompareRunTableToolbar: React.FC<FilterProps> = ({ ...toolbarProps }) => {
           Manage runs
         </Button>
       </ToolbarItem>
-    </PipelineFilterBar>
+    </PipelineRunTableToolbarBase>
   );
 };
