+++ frontend/src/concepts/pipelines/content/tables/pipelineRun/PipelineRunTableToolbar.tsx
@@ -1,23 +1,12 @@
 import * as React from 'react';
-import { TextInput, ToolbarItem } from '@patternfly/react-core';
+import { ToolbarItem } from '@patternfly/react-core';
 import { useParams } from 'react-router-dom';
-import PipelineFilterBar from '~/concepts/pipelines/content/tables/PipelineFilterBar';
-import SimpleSelect from '~/components/SimpleSelect';
 import { FilterOptions } from '~/concepts/pipelines/content/tables/usePipelineFilter';
-import ExperimentSearchInput from '~/concepts/pipelines/content/tables/ExperimentSearchInput';
-import { RuntimeStateKF, runtimeStateLabels } from '~/concepts/pipelines/kfTypes';
-import DashboardDatePicker from '~/components/DashboardDatePicker';
-import PipelineVersionSelect from '~/concepts/pipelines/content/pipelineSelector/CustomPipelineVersionSelect';
-import { PipelineRunVersionsContext } from '~/pages/pipelines/global/runs/PipelineRunVersionsContext';
-
-export type FilterProps = Pick<
-  React.ComponentProps<typeof PipelineFilterBar>,
-  'filterData' | 'onFilterUpdate'
->;
+import PipelineRunTableToolbarBase, { FilterProps } from './PipelineRunTableToolbarBase';
 
 interface PipelineRunTableToolbarProps extends FilterProps {
   actions?: React.ReactNode[];
-  filterOptions?: React.ComponentProps<typeof PipelineFilterBar>['filterOptions'];
+  filterOptions?: React.ComponentProps<typeof PipelineRunTableToolbarBase>['filterOptions'];
 }
 
 const PipelineRunTableToolbar: React.FC<PipelineRunTableToolbarProps> = ({
@@ -25,16 +14,7 @@ const PipelineRunTableToolbar: React.FC<PipelineRunTableToolbarProps> = ({
   filterOptions,
   ...toolbarProps
 }) => {
-  const { versions } = React.useContext(PipelineRunVersionsContext);
   const { experimentId, pipelineVersionId } = useParams();
-  /* eslint-disable @typescript-eslint/no-unused-vars */
-  const {
-    [RuntimeStateKF.RUNTIME_STATE_UNSPECIFIED]: unspecifiedState,
-    [RuntimeStateKF.PAUSED]: pausedState,
-    [RuntimeStateKF.CANCELED]: cancelledState,
-    ...statusRuntimeStates
-  } = runtimeStateLabels;
-  /* eslint-enable @typescript-eslint/no-unused-vars */
 
   const defaultFilterOptions = React.useMemo(
     () => ({
@@ -52,63 +32,13 @@ const PipelineRunTableToolbar: React.FC<PipelineRunTableToolbarProps> = ({
   );
 
   return (
-    <PipelineFilterBar
+    <PipelineRunTableToolbarBase
       {...toolbarProps}
       filterOptions={filterOptions || defaultFilterOptions}
-      filterOptionRenders={{
-        [FilterOptions.NAME]: ({ onChange, ...props }) => (
-          <TextInput
-            {...props}
-            data-testid="search-for-run-name"
-            aria-label="Search for a run name"
-            placeholder="Search..."
-            onChange={(_event, value) => onChange(value)}
-          />
-        ),
-        [FilterOptions.EXPERIMENT]: ({ onChange, value, label }) => (
-          <ExperimentSearchInput
-            onChange={(data) => onChange(data?.value, data?.label)}
-            selected={value && label ? { value, label } : undefined}
-          />
-        ),
-        [FilterOptions.PIPELINE_VERSION]: ({ onChange, label }) => (
-          <PipelineVersionSelect
-            versions={versions}
-            selection={label}
-            onSelect={(version) => onChange(version.pipeline_version_id, version.display_name)}
-          />
-        ),
-        [FilterOptions.CREATED_AT]: ({ onChange, ...props }) => (
-          <DashboardDatePicker
-            {...props}
-            hideError
-            aria-label="Select a start date"
-            onChange={(_, value, date) => {
-              if (date || !value) {
-                onChange(value);
-              }
-            }}
-          />
-        ),
-        [FilterOptions.STATUS]: ({ value, onChange, ...props }) => (
-          <SimpleSelect
-            {...props}
-            value={value ?? ''}
-            aria-label="Select a status"
-            options={Object.values(statusRuntimeStates).map((v) => ({
-              key: v,
-              label: v,
-            }))}
-            onChange={(v) => onChange(v)}
-            dataTestId="runtime-status-dropdown"
-          />
-        ),
-      }}
-    >
-      {actions?.map((action, index) => (
+      actions={actions?.map((action, index) => (
         <ToolbarItem key={index}>{action}</ToolbarItem>
       ))}
-    </PipelineFilterBar>
+    />
   );
 };
